\name{test.data}
\alias{test.data}
\docType{data}
\title{
Example test data
}
\description{
This dataset contains binary responses of 775 participants to 20 items.Missing responses are coded as NA.
}
\usage{data("test.data")}
\format{
  A data frame with 775 observations on the following 20 items.
  \describe{
    \item{\code{Item_1}}{a numeric vector}
    \item{\code{Item_2}}{a numeric vector}
    \item{\code{Item_3}}{a numeric vector}
    \item{\code{Item_4}}{a numeric vector}
    \item{\code{Item_5}}{a numeric vector}
    \item{\code{Item_6}}{a numeric vector}
    \item{\code{Item_7}}{a numeric vector}
    \item{\code{Item_8}}{a numeric vector}
    \item{\code{Item_9}}{a numeric vector}
    \item{\code{Item_10}}{a numeric vector}
    \item{\code{Item_11}}{a numeric vector}
    \item{\code{Item_12}}{a numeric vector}
    \item{\code{Item_13}}{a numeric vector}
    \item{\code{Item_14}}{a numeric vector}
    \item{\code{Item_15}}{a numeric vector}
    \item{\code{Item_16}}{a numeric vector}
    \item{\code{Item_17}}{a numeric vector}
    \item{\code{Item_18}}{a numeric vector}
    \item{\code{Item_19}}{a numeric vector}
    \item{\code{Item_20}}{a numeric vector}
  }
}
\details{
A test data that contain binary responses of 775 participants to 20 items.Missing responses are coded as NA.
}
\examples{
data(test.data)
## maybe str(test.data) ; plot(test.data) ...
}
\keyword{datasets}
