% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listwise.r
\name{Listwise}
\alias{Listwise}
\title{Listwise Deletion (LW)}
\usage{
Listwise(test.data, Mvalue = "NA")
}
\arguments{
\item{test.data}{Test data set (a data frame or a matrix) containing missing responses. 
Missing values are coded as NA or other values (e.g., 8, 9).#' @param Mvalue  Missing response indicators in the data (e.g. "NA", "8", "9", etc.). Mvalue="NA" by default.}

\item{Mvalue}{Missing response indicators in the data (e.g. "NA", "8", "9", etc.). Mvalue="NA" by default.}
}
\value{
A data frame with no missing responses.
}
\description{
This function deletes examinees who report missing responses.
}
\examples{
 
        Listwise(test.data, Mvalue="8")
}

