% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatIncorrect.r
\name{TreatIncorrect}
\alias{TreatIncorrect}
\title{Treat missing responses as incorrect (IN)}
\usage{
TreatIncorrect(test.data, Mvalue = "NA")
}
\arguments{
\item{test.data}{Test data set (a data frame or a matrix) containing missing responses. 
Missing values are coded as NA or other values (e.g., 8, 9).}

\item{Mvalue}{Missing response indicators in the data (e.g. "NA", "8", "9", etc.). Mvalue="NA" by default.}
}
\value{
A data frame with all missing responses replaced by '0'.
}
\description{
This function replaces all missing responses by zero (see Lord, 1974 <doi: 10.1111/j.1745-3984.1974.tb00996.x>; 
 Mislevy & Wu, 1996 <doi: 10.1002/j.2333-8504.1996.tb01708.x>; Pohl et al., 2014 <doi: 10.1177/0013164413504926>);).
}
\examples{
 
        TreatIncorrect(test.data, Mvalue="NA")
}
\references{
{
 Lord, F. M. (1974).
" Quick estimates of the relative efficiency of two tests as a function of ability level."
 Journal of Educational Measurement, 11(4), 247-254. doi: 10.1111/j.1745-3984.1974.tb00996.x. 
}

{
 Mislevy, R. J., & Wu, P. K. (1996).
" Missing responses and IRT ability estimation: Omits, choice, time limits, and adaptive testing. "
 ETS Research Report Series, 1996(2), i-36. doi: 10.1002/j.2333-8504.1996.tb01708.x. 
}

{
Pohl, S., Gräfe, L., & Rose, N. (2014).
"Dealing with omitted and not-reached items in competence tests evaluating approaches accounting for missing responses in item response theory models. "
Educational and Psychological Measurement, 74(3), 423-452. doi: 10.1177/0013164413504926. 
}
}
